Pull metrics test out to a separate file. Make jaeger test use new assertion of reload rather than waiting on 100 millis
