+++ apollo-router/tests/fixtures/jaeger.router.yaml
@@ -1,9 +1,4 @@
 telemetry:
-  metrics:
-    prometheus:
-      listen: 127.0.0.1:4000
-      enabled: true
-      path: /metrics
   tracing:
     experimental_response_trace_id:
       enabled: true
