+++ apollo-router/tests/fixtures/prometheus.router.yaml
@@ -1,0 +1,10 @@
+telemetry:
+  metrics:
+    prometheus:
+      listen: 127.0.0.1:4000
+      enabled: true
+      path: /metrics
+override_subgraph_url:
+  products: http://localhost:4005
+include_subgraph_errors:
+  all: true
