+++ apollo-router/tests/jaeger_test.rs
@@ -14,7 +14,7 @@ use crate::common::IntegrationTest;
 use crate::common::ValueExt;
 
 #[tokio::test(flavor = "multi_thread")]
-async fn test_jaeger_tracing_and_metrics() -> Result<(), BoxError> {
+async fn test_jaeger_tracing() -> Result<(), BoxError> {
     let tracer = opentelemetry_jaeger::new_agent_pipeline()
         .with_service_name("my_app")
         .install_simple()?;
@@ -29,7 +29,7 @@ async fn test_jaeger_tracing_and_metrics() -> Result<(), BoxError> {
     router.start().await;
     router.assert_started().await;
 
-    for _ in 0..10 {
+    for _ in 0..2 {
         let (id, result) = router.run_query().await;
         assert!(!result
             .headers()
@@ -38,17 +38,9 @@ async fn test_jaeger_tracing_and_metrics() -> Result<(), BoxError> {
             .is_empty());
         query_jaeger_for_trace(id).await?;
         router.touch_config().await;
-        tokio::time::sleep(Duration::from_millis(100)).await;
+        router.assert_reloaded().await;
     }
-
-    let metrics = router.get_metrics().await.unwrap();
-    assert!(metrics.contains(r#"apollo_router_cache_hit_count{kind="query planner",service_name="apollo-router",storage="memory"} 9"#));
-    assert!(metrics.contains(r#"apollo_router_cache_miss_count{kind="query planner",service_name="apollo-router",storage="memory"} 1"#));
-    assert!(metrics.contains("apollo_router_cache_hit_time"));
-    assert!(metrics.contains("apollo_router_cache_miss_time"));
-    assert!(metrics.contains("apollo_router_session_count_total"));
-    assert!(metrics.contains("apollo_router_session_count_active"));
-
+    router.graceful_shutdown().await;
     Ok(())
 }
